// src/api/generated/datakatalog/zod.gen.ts
export const zAssosiasjonstype = z.object({
    id: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    navn: z.optional(z.string()),
    egenskapstype: z.string(),
    maksimalt_antall_verdier: z.optional(z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' })),
    minimalt_antall_verdier: z.optional(z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' })),
    innhold: z.optional(zAssosiasjonstypeListeInnhold),
    type: z.optional(zAssosiasjonstypeVegobjekttype),
    innenfor_mor: z.optional(z.enum([
        'NEI',
        'JA',
        'MED_AVVIK'
    ])),
    relasjonstype: z.optional(z.enum([
        'ASSOSIASJON',
        'AGGREGERING',
        'KOMPOSISJON',
        'TOPOLOGI'
    ]))
});

export const zEgenskapstypeEnum = z.enum([
    'Assosiasjon',
    'Boolsk',
    'Binær',
    'Tekst',
    'Dato',
    'Flyttall',
    'Heltall',
    'Struktur',
    'Geometri',
    'Stedfesting',
    'Kortdato',
    'Tid',
    'Liste',
    'Tekstenum',
    'Heltallenum',
    'Flyttallenum'
]);
// ...
export const zRelasjonstyper = z.object({
    foreldre: z.array(zAssosiasjonstype),
    barn: z.array(zAssosiasjonstype)
});

export const zVegobjektKategori = z.object({
    id: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    primærkategori: z.boolean(),
    nummer: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    navn: z.optional(z.string()),
    kortnavn: z.optional(z.string()),
    sorteringsnummer: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    beskrivelse: z.optional(z.string()),
    startDato: z.optional(z.iso.date())
});

export const zViktighet = z.enum([
    'IKKE_SATT',
    'PÅKREVD_ABSOLUTT',
    'PÅKREVD_IKKE_ABSOLUTT',
    'BETINGET',
    'OPSJONELL',
    'MINDRE_VIKTIG',
    'HISTORISK'
]);

export const zEgenskapstype = z.object({
    id: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    navn: z.optional(z.string()),
    egenskapstype: zEgenskapstypeEnum,
    komplementær_egenskapstype: z.optional(z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' })),
    kortnavn: z.optional(z.string()),
    beskrivelse: z.optional(z.string()),
    veiledning: z.optional(z.string()),
    sosinavn: z.optional(z.string()),
    sosinvdbnavn: z.optional(z.string()),
    sorteringsnummer: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    avledet: z.boolean(),
    sensitivitet: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    skrivebeskyttet: z.boolean(),
    obligatorisk_verdi: z.boolean(),
    gruppesorteringsnummer: z.optional(z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' })),
    grunnrissreferanse: z.optional(z.string()),
    høydereferanse: z.optional(z.string()),
    høydereferanse_tall: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    nøyaktighetskrav_grunnriss: z.number(),
    nøyaktighetskrav_høyde: z.number(),
    sosi_referanse: z.optional(z.string()),
    referansegeometri_tilstrekkelig: z.boolean(),
    viktighet: zViktighet,
    kategori: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    tilleggskrav: z.optional(z.array(z.string()))
});
// ...
export const zStedfesting = z.object({
    id: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    navn: z.optional(z.string()),
    egenskapstype: zEgenskapstypeEnum,
    komplementær_egenskapstype: z.optional(z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' })),
    kortnavn: z.optional(z.string()),
    beskrivelse: z.optional(z.string()),
    veiledning: z.optional(z.string()),
    sosinavn: z.optional(z.string()),
    sosinvdbnavn: z.optional(z.string()),
    sorteringsnummer: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    avledet: z.boolean(),
    sensitivitet: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    skrivebeskyttet: z.boolean(),
    obligatorisk_verdi: z.boolean(),
    gruppesorteringsnummer: z.optional(z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' })),
    grunnrissreferanse: z.optional(z.string()),
    høydereferanse: z.optional(z.string()),
    høydereferanse_tall: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    nøyaktighetskrav_grunnriss: z.number(),
    nøyaktighetskrav_høyde: z.number(),
    sosi_referanse: z.optional(z.string()),
    referansegeometri_tilstrekkelig: z.boolean(),
    viktighet: zViktighet,
    kategori: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' })
});
// ...
export const zVegobjekttype = z.object({
    id: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    navn: z.optional(z.string()),
    kortnavn: z.optional(z.string()),
    beskrivelse: z.optional(z.string()),
    objektliste_dato: z.optional(z.iso.date()),
    veiledning: z.optional(z.string()),
    stedfesting: z.optional(zStedfesting),
    sosinavn: z.optional(z.string()),
    sosinvdbnavn: z.optional(z.string()),
    sorteringsnummer: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    status: z.optional(z.string()),
    hovedkategori: z.optional(z.string()),
    tidsrom_relevant: z.boolean(),
    konnekteringslenke_ok: z.optional(z.boolean()),
    abstrakt_type: z.boolean(),
    avledet: z.boolean(),
    må_ha_mor: z.boolean(),
    er_dataserie: z.boolean(),
    en_versjon: z.boolean(),
    tilleggsinformasjon: z.optional(z.string()),
    kategorier: z.array(zVegobjektKategori),
    egenskapstyper: z.optional(z.array(zEgenskapstype)),
    relasjonstyper: z.optional(zRelasjonstyper),
    sensitiv: z.boolean()
});